requirements.txt
anyio==4.3.0
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
distlib==0.3.8
Django==5.0.6
Everything-Tkinter==0.2
filelock==3.14.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
jalali_core==1.0.0
jdatetime==5.0.0
platformdirs==4.2.2
PyMySQL==1.1.0
PyQt6==6.6.1
PyQt6-Qt6==6.6.3
PyQt6-sip==13.6.0
python-telegram-bot==21.0.1
request2==0.2
requests==2.31.0
sniffio==1.3.1
SQLite3-0611==0.0.1
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.2
mysql-connector-python

